+++ Assassin/Configurations.cpp
@@ -1,5 +1,5 @@
-#include "pch.h"
-#include "NagisaCore.Configuration.h"
+﻿#include "pch.h"
+#include "Configurations.h"
 
 using namespace concurrency;
 using namespace Platform;
@@ -7,8 +7,7 @@ using namespace Windows::Foundation;
 using namespace Windows::Storage;
 using namespace Windows::Storage::AccessCache;
 
-
-NagisaCore::Configuration::Configuration()
+Assassin::Configurations::Configurations()
 {
 	m_FutureAccessList = StorageApplicationPermissions::FutureAccessList;
 	//m_LocalSettings = ApplicationData::Current->LocalSettings;
@@ -16,12 +15,10 @@ NagisaCore::Configuration::Configuration()
 	m_DownloadsFolder = GetDownloadsFolder();
 }
 
-
-
-StorageFolder^ NagisaCore::Configuration::GetDownloadsFolder()
+StorageFolder^ Assassin::Configurations::GetDownloadsFolder()
 {
 	StorageFolder^ Folder = nullptr;
-	
+
 	if (m_FutureAccessList->ContainsItem(L"Nagisa.DownloadsFolder"))
 	{
 		Folder = m2_await(
@@ -31,13 +28,29 @@ StorageFolder^ NagisaCore::Configuration::GetDownloadsFolder()
 	return Folder;
 }
 
-StorageFolder^ NagisaCore::Configuration::DownloadsFolder::get()
+StorageFolder^ Assassin::Configurations::DownloadsFolder::get()
 {
 	return m_DownloadsFolder;
 }
 
-void NagisaCore::Configuration::DownloadsFolder::set(StorageFolder^ value)
+void Assassin::Configurations::DownloadsFolder::set(StorageFolder^ value)
 {
 	m_FutureAccessList->AddOrReplace(L"Nagisa.DownloadsFolder", value);
 	m_DownloadsFolder = value;
+}
+
+String^ Assassin::Configurations::Version::get()
+{
+	return NAGISA_VERSION_STRING;
+}
+
+String^ Assassin::Configurations::UserAgent::get()
+{
+	return 
+		L"Mozilla/5.0 "
+		L"(Windows NT 10.0; Nagisa/" NAGISA_VERSION_STRING L") "
+		L"AppleWebKit/537.36 (KHTML, like Gecko) "
+		L"Chrome/42.0.2311.135 "
+		L"Safari/537.36 "
+		L"Edge/12.10240";
 }